+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.14] - 2024-10-10
+
+### Added
+- Add convenience `MoodleContentHandler` model
+
 ## [0.1.13] - 2024-10-10
 
 ### Added